+++ include/PowerPC_EABI_Support/MSL_C/MSL_Common/arith.h
@@ -13,7 +13,7 @@ typedef struct {
 } div_t;
 
 int abs(int __x);
-long labs(long __x);
+s32 labs(s32 __x);
 div_t div(s32 __numer, s32 __denom);
 
 #ifdef __cplusplus
